+++ whats4/src/App.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import './App.css';
-// import InputBotao from './components/InputBotao/InputBotao';
 import ListaMensagem from './components/ListaMensagem/ListaMensagem';
