+++ whats4/src/components/ListaMensagem/ListaMensagem.js
@@ -1,7 +1,5 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import InputBotao from '../InputBotao/InputBotao'
-
 
 class ListaMensagem extends React.Component {
     constructor (props) {
@@ -9,26 +7,66 @@ class ListaMensagem extends React.Component {
 
         this.state = {
             arrayMensagem: [
-                {
-                    nome: "",
-                    mensagem: ""
-                }
-            ]
+                
+            ],
+
+            valorInputNome: "",
+            valorInputMensagem: ""
+        };
+    }
+    
+    enviaMensagem = () => {
+        const novaMensagem = {
+            nome: this.state.valorInputNome,
+            mensagem: this.state.valorInputMensagem
         }
+
+        const novasMensagens = [
+            novaMensagem, ...this.state.arrayMensagem
+        ]
+
+        this.setState({
+            arrayMensagem: novasMensagens,
+            valorInputNome: "",
+            valorInputMensagem: ""
+        })
     }
 
-    componenteMensagem = <InputBotao
-        
-    />
+    onChangeInputNome = event => {
+        this.setState ({ valorInputNome: event.target.value })
+    }
+    
+    onChangeInputMensagem = event => {
+        this.setState ({ valorInputMensagem: event.target.value })
+    }
+
     
     render() {
+        const containerDasMensagens =
+        this.state.arrayMensagem.map(mensagem => {
+            return (
+                <p>
+                    <strong>{mensagem.nome+":"}</strong> {mensagem.mensagem}
+                </p>
+
+            )
+        })
+    
         return (
             <div>
-                <p>{ this.state.nome, this.state.mensagem }</p>
+                <div>{ containerDasMensagens }</div>
 
-                <InputBotao 
-                    
+                <input 
+                    value={this.state.valorInputNome} 
+                    onChange={this.onChangeInputNome}
+                    placeholder="Nome"
                 />
+                <input 
+                    value={this.state.valorInputMensagem}
+                    onChange={this.onChangeInputMensagem}
+                    placeholder="Digite sua mensagem"
+                />
+                <button onClick={this.enviaMensagem}>Enviar</button>
             </div>
         )
     }
